# src/api/api.yaml
openapi: 3.0.0
paths:
...
  /server/commands/all:
    get:
      operationId: ServerCommandsController_getCommands
      parameters:
        - name: categoryId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandListResponseDto'
      tags: &ref_8
        - Server
      security:
        - bearer: []
  /server/commands/{id}:
    patch:
      operationId: ServerCommandsController_putCommand
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchServerCommandDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponseDto'
      tags: *ref_8
      security:
        - bearer: []
  /server/commands/send/{id}:
    post:
      operationId: ServerCommandsController_startServer
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandExecuteResponseDto'
      tags: *ref_8
      security:
        - bearer: []
  /server/commands/{id}/markers:
    get:
      operationId: ServerCommandsController_getProgressMarkers
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandProgressMarkerListResponseDto'
      tags: *ref_8
      security:
        - bearer: []
    post:
      operationId: ServerCommandsController_createProgressMarker
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommandProgressMarkerDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandProgressMarkerResponseDto'
      tags: *ref_8
      security:
        - bearer: []
  /server/commands/markers/{markerId}:
    patch:
      operationId: ServerCommandsController_updateProgressMarker
      parameters:
        - name: markerId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommandProgressMarkerDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandProgressMarkerResponseDto'
      tags: *ref_8
      security:
        - bearer: []
    delete:
      operationId: ServerCommandsController_deleteProgressMarker
      parameters:
        - name: markerId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandProgressMarkerResponseDto'
      tags: *ref_8
      security:
        - bearer: []
  /server/process/all:
    get:
      operationId: ServerProcessController_getAll
      parameters:
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessListResponseDto'
      tags: &ref_9
        - Server
      security:
        - bearer: []
  /server/process/{id}:
    get:
      operationId: ServerProcessController_getOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessResponseDto'
      tags: *ref_9
      security:
        - bearer: []
    delete:
      operationId: ServerProcessController_delete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Process deleted
      tags: *ref_9
      security:
        - bearer: []
  /server/process/{id}/logs:
    get:
      operationId: ServerProcessController_getAllLogs
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessLogListResponseDto'
      tags: *ref_9
      security:
        - bearer: []
...
  /server/all:
    get:
      operationId: ServerController_getAll
      parameters:
        - name: take
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 20
            type: number
        - name: skip
          required: false
          in: query
          schema:
            minimum: 0
            default: 0
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerListResponseDto'
      tags: &ref_11
        - Server
      security:
        - bearer: []
  /server/{serverId}:
    get:
      operationId: ServerController_get
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponseDto'
      tags: *ref_11
      security:
        - bearer: []
  /server/details/{serverId}:
    get:
      operationId: ServerController_getDetails
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerDetailsResponseDto'
      tags: *ref_11
      security:
        - bearer: []
  /server/{serverId}/start:
    patch:
      operationId: ServerController_start
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PowerServerResponseDto'
      tags: *ref_11
      security:
        - bearer: []
  /server/{serverId}/stop:
    patch:
      operationId: ServerController_stop
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PowerServerResponseDto'
      tags: *ref_11
      security:
        - bearer: []
  /server/{serverId}/reboot:
    patch:
      operationId: ServerController_reboot
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PowerServerResponseDto'
      tags: *ref_11
      security:
        - bearer: []
  /server/disk/{id}:
    patch:
      operationId: ServerController_patchDisk
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDiskDto'
      responses:
        '200':
          description: Changed correctly
      tags: *ref_11
      security:
        - bearer: []
  /server/{serverId}/category:
    post:
      operationId: ServerController_createCategory
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '200':
          description: Changed correctly
      tags: *ref_11
      security:
        - bearer: []
  /server/category/{id}:
    patch:
      operationId: ServerController_patchCategory
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchCategoryDto'
      responses:
        '200':
          description: Changed correctly
      tags: *ref_11
      security:
        - bearer: []
...
info:
  title: WHCP Backend
  description: XD
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
    JWT-refresh:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT refresh
      description: Enter JWT refresh token
      in: header
    JWT-reset-password:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT reset password
      description: Enter JWT reset password token
      in: header
    JWT-register-user:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT register user
      description: Enter JWT register user
      in: header
  schemas:
...
    PaginationDto:
      type: object
      properties:
        take:
          type: number
          minimum: 1
          maximum: 20
        skip:
          type: number
          minimum: 0
          default: 0
...
    Role:
      type: string
      enum:
        - USER
        - MODERATOR
        - ADMIN
        - OWNER
    UserDto:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: '#/components/schemas/Role'
      required:
        - email
        - role
...
    CommandStatus:
      type: string
      enum:
        - ENABLED
        - DISABLED
        - RUNNING
    CommandType:
      type: string
      enum:
        - MESSAGE
        - EVENT
    CommandResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value:
          type: string
        status:
          $ref: '#/components/schemas/CommandStatus'
        type:
          $ref: '#/components/schemas/CommandType'
      required:
        - id
        - value
        - status
        - type
    CommandListResponseDto:
      type: object
      properties:
        commands:
          type: array
          items:
            $ref: '#/components/schemas/CommandResponseDto'
        total:
          type: number
      required:
        - commands
        - total
    PatchServerCommandDto:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    CommandExecuteResponseDto:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    CommandMatchType:
      type: string
      enum:
        - CONTAINS
        - STARTS_WITH
        - REGEX
    CommandRuntimeStatus:
      type: string
      enum:
        - IDLE
        - STARTING
        - RUNNING
        - STOPPING
        - STOPPED
        - ERROR
    ProcessLogLevel:
      type: string
      enum:
        - LOG
        - SUCCESS
        - WARNING
        - ERROR
    CommandProgressMarkerResponseDto:
      type: object
      properties:
        id:
          type: string
        commandId:
          type: string
        label:
          type: string
        pattern:
          type: string
        matchType:
          $ref: '#/components/schemas/CommandMatchType'
        progress:
          type: number
        runtimeStatus:
          $ref: '#/components/schemas/CommandRuntimeStatus'
        level:
          $ref: '#/components/schemas/ProcessLogLevel'
        order:
          type: number
      required:
        - id
        - commandId
        - pattern
        - matchType
        - order
    CommandProgressMarkerListResponseDto:
      type: object
      properties:
        markers:
          type: array
          items:
            $ref: '#/components/schemas/CommandProgressMarkerResponseDto'
        total:
          type: number
      required:
        - markers
        - total
    CreateCommandProgressMarkerDto:
      type: object
      properties:
        label:
          type: string
        pattern:
          type: string
        matchType:
          $ref: '#/components/schemas/CommandMatchType'
        progress:
          type: number
          minimum: 0
          maximum: 100
        runtimeStatus:
          $ref: '#/components/schemas/CommandRuntimeStatus'
        level:
          $ref: '#/components/schemas/ProcessLogLevel'
        order:
          type: number
          minimum: 0
      required:
        - pattern
    UpdateCommandProgressMarkerDto:
      type: object
      properties:
        label:
          type: string
        pattern:
          type: string
        matchType:
          $ref: '#/components/schemas/CommandMatchType'
        progress:
          type: number
          minimum: 0
          maximum: 100
        runtimeStatus:
          $ref: '#/components/schemas/CommandRuntimeStatus'
        level:
          $ref: '#/components/schemas/ProcessLogLevel'
        order:
          type: number
          minimum: 0
    ServerProcessStatus:
      type: string
      enum:
        - UNKNOWN
        - STARTED
        - ONGOING
        - CLOSED
        - ENDED
        - FAILED
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
      required:
        - id
        - email
    Category:
      type: object
      properties:
        name:
          type: string
    ProcessResponseDto:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/ServerProcessStatus'
        runtimeStatus:
          $ref: '#/components/schemas/CommandRuntimeStatus'
        progress:
          type: number
        name:
          type: string
        startedAt:
          type: string
          format: date-time
        stoppedAt:
          type: string
          format: date-time
        startedBy:
          $ref: '#/components/schemas/User'
        category:
          $ref: '#/components/schemas/Category'
      required:
        - id
        - status
        - runtimeStatus
        - startedAt
    ProcessListResponseDto:
      type: object
      properties:
        processes:
          type: array
          items:
            $ref: '#/components/schemas/ProcessResponseDto'
        total:
          type: number
        params:
          $ref: '#/components/schemas/PaginationDto'
      required:
        - processes
        - total
    ProcessLog:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        level:
          $ref: '#/components/schemas/ProcessLogLevel'
      required:
        - id
        - message
        - timestamp
        - level
    ProcessLogListResponseDto:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/ProcessLog'
        total:
          type: number
        params:
          $ref: '#/components/schemas/PaginationDto'
      required:
        - logs
        - total
...
    ServerResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ipAddress:
          type: string
        macAddress:
          type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - ipAddress
        - createdAt
        - updatedAt
    ServerListResponseDto:
      type: object
      properties:
        servers:
          type: array
          items:
            $ref: '#/components/schemas/ServerResponseDto'
        total:
          type: number
        params:
          $ref: '#/components/schemas/PaginationDto'
      required:
        - servers
        - total
    ServerStatus:
      type: string
      enum:
        - UNKNOWN
        - OFFLINE
        - ONLINE
        - ERROR
        - MAINTENANCE
        - WAKE_IN_PROGRESS
        - SHUTDOWN_IN_PROGRESS
    CpuDto:
      type: object
      properties:
        cores:
          type: number
        physicalCores:
          type: number
        currentLoad:
          type: number
        currentLoadUser:
          type: number
        currentLoadSystem:
          type: number
      required:
        - currentLoad
        - currentLoadUser
        - currentLoadSystem
    MemoryDto:
      type: object
      properties:
        total:
          type: number
        free:
          type: number
    DiskType:
      type: string
      enum:
        - HDD
        - SSD
        - NVME
    DiskInfoDto:
      type: object
      properties:
        fs:
          type: string
        type:
          type: string
        used:
          type: number
        available:
          type: number
        name:
          type: string
        mediaType:
          $ref: '#/components/schemas/DiskType'
    ServerPropertiesDto:
      type: object
      properties:
        uptime:
          type: number
        status:
          $ref: '#/components/schemas/ServerStatus'
        lastSeenAt:
          type: string
          format: date-time
        isOnline:
          type: boolean
        startedBy:
          $ref: '#/components/schemas/UserDto'
        cpuInfo:
          $ref: '#/components/schemas/CpuDto'
        memoryInfo:
          $ref: '#/components/schemas/MemoryDto'
        diskInfo:
          type: array
          items:
            $ref: '#/components/schemas/DiskInfoDto'
    ServerCategoriesDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value:
          type: string
      required:
        - id
        - value
    ServerDetailsResponseDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ipAddress:
          type: string
        location:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        properties:
          $ref: '#/components/schemas/ServerPropertiesDto'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/ServerCategoriesDto'
      required:
        - id
        - name
        - ipAddress
        - createdAt
    PowerServerResponseDto:
      type: object
      properties:
        success:
          type: boolean
        serverId:
          type: string
        newStatus:
          $ref: '#/components/schemas/ServerStatus'
        message:
          type: string
      required:
        - success
        - serverId
        - newStatus
    PatchDiskDto:
      type: object
      properties:
        name:
          type: string
        mediaType:
          $ref: '#/components/schemas/DiskType'
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
      required:
        - value
    PatchCategoryDto:
      type: object
      properties:
        name:
          type: string
...